RDCC-3511:  Changes made to code reformating
