+++ src/main/java/uk/gov/hmcts/reform/judicialapi/controller/TestingSupportController.java
@@ -10,7 +10,7 @@ import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.reform.judicialapi.service.JudicialUserService;
+import uk.gov.hmcts.reform.judicialapi.service.IdamUserProfileService;
 
 
 @RestController
@@ -25,11 +25,11 @@ import uk.gov.hmcts.reform.judicialapi.service.JudicialUserService;
 public class TestingSupportController {
 
     @Autowired
-    JudicialUserService judicialUserService;
+    IdamUserProfileService idamUserProfileService;
 
     @GetMapping(path = "/testing-support/sidam/actions/create-users", produces = "application/json")
     @Secured({"jrd-system-user", "jrd-admin"})
     public ResponseEntity<Object> createIdamUserProfiles() {
-        return judicialUserService.createIdamUserProfiles();
+        return idamUserProfileService.createIdamUserProfiles();
     }
 }
