+++ src/main/java/uk/gov/hmcts/reform/judicialapi/repository/UserProfileRepository.java
@@ -35,6 +35,4 @@ public interface UserProfileRepository extends JpaRepository<UserProfile, String
     List<UserProfile> findBySearchString(String searchString, String serviceCode, String locationCode,
                                          List<String> ticketCode);
 
-    @Override
-    List<UserProfile> findAll();
 }
