+++ src/main/java/uk/gov/hmcts/reform/judicialapi/service/IdamUserProfileService.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.reform.judicialapi.service;
+
+import org.springframework.http.ResponseEntity;
+
+public interface IdamUserProfileService {
+    public ResponseEntity<Object> createIdamUserProfiles();
+}
