+++ src/main/java/uk/gov/hmcts/reform/judicialapi/util/JrdConstant.java
@@ -7,6 +7,4 @@ public final class JrdConstant {
 
     public static final String USER_DATA_NOT_FOUND = "User data not found for the request provided";
 
-    public static final String IDAM_USER_CREATED_SUCCESS = "Idam user profile create successfully";
-    public static final String IDAM_USER_CREATED_FAIL = "Idam user creation failed";
 }
